Alteração da aula dia 08/09/2025

+++ biblioteca/src/main/java/br/com/fuctura/biblioteca/services/CategoriaService.java
@@ -1,8 +1,10 @@
 package br.com.fuctura.biblioteca.services;
 
+import br.com.fuctura.biblioteca.exeptions.ObjectNotFoundExepiton;
 import br.com.fuctura.biblioteca.models.Categoria;
 import br.com.fuctura.biblioteca.repositores.CategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
@@ -17,7 +19,10 @@ public class CategoriaService {
 
     public Categoria findById(Integer id) {
         Optional<Categoria> categoria = categoriaRepository.findById(id);
-        return categoria.orElse(null);
+        if (categoria.isPresent()) {
+            return categoria.get();
+        }
+        throw new ObjectNotFoundExepiton("Categorai Não encontrada ou não existe");
     }
 
     public List<Categoria> findAll() {
@@ -25,16 +30,31 @@ public class CategoriaService {
     }
 
     public Categoria save(Categoria categoria) {
-       Categoria cat = categoriaRepository.save(categoria);
-       return cat;
+        findByGenero(categoria);
+        Categoria cat = categoriaRepository.save(categoria);
+        return cat;
     }
 
     public Categoria update(@RequestBody Categoria categoria) {
+        findById(categoria.getId());
+        findByGenero(categoria);
         Categoria cat = categoriaRepository.save(categoria);
         return cat;
     }
 
     public void delete(Integer id) {
+        Categoria cat = findById(id);
+        if (!cat.getLivros().isEmpty()) {
+
+            throw new DataIntegrityViolationException("Categoria possui livros, não pode ser deletada");
+        }
         categoriaRepository.deleteById(id);
+    }
+
+    public void findByGenero(Categoria categoria) {
+        Optional<Categoria> cat = categoriaRepository.findByGenero(categoria.getGenero());
+        if (cat.isPresent() && cat.get().getGenero().equals(categoria.getGenero())) {
+            throw new IllegalArgumentException("Já existe uma categoria cadastrada com este genero");
+        }
     }
 }
